+++ CHANGELOG.md
@@ -4,9 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Create `ExportIcon` component.
-- Feature: Update `EmailIcon` component.
-- Feature: Add `forceOneLine` prop to `TextCopy` component.
+## [2.17.0] - 2019-07-24
+
+Features:
+- Create `ExportIcon` component.
+- Update `EmailIcon` component.
+- Add `forceOneLine` prop to `TextCopy` component.
 
 ## [2.16.2] - 2019-07-19